+++ local-s3-rest/src/main/java/com/robothy/s3/rest/handler/ListObjectsV2Controller.java
@@ -43,6 +43,8 @@ public class ListObjectsV2Controller implements HttpRequestHandler {
         boolean fetchOwner = request.parameter("fetch-owner").map(Boolean::valueOf).orElse(false);
         ListObjectsV2Ans listObjectsV2Ans = this.objectService.listObjectsV2(bucket, continuationToken, delimiter, encodingType, fetchOwner, maxKeys, prefix, startAfter);
 
+        OptionalObjectAttributes.apply(request, listObjectsV2Ans.getObjects());
+
         ListBucketV2Result listBucketV2Result = ListBucketV2Result.builder()
             .isTruncated(listObjectsV2Ans.isTruncated())
             .contents(listObjectsV2Ans.getObjects())